NETCODE CHANGE: No longer using network ID, use steamid instead

+++ Assets/codes/Network/Packets/ClientSend.cs
@@ -77,7 +77,7 @@ public class ClientSend
     }
 
 
-    public static Result PickUpItem(string objectID, int whopicked)
+    public static Result PickUpItem(string objectID, ulong whopicked)
     {
         using (packet p = new packet((int)ClientPackets.PickUpItem))
         {
